pyproject.toml
[tool.poetry]
name = "sql-app"
version = "0.1.0"
description = ""
authors = ["Brandon Severin <brandon.severin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.23"
pydantic = "^2.5.2"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
